+++ CHANGELOG.rst
@@ -4,6 +4,14 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+1.2.1
+-----
+
+Fixed
+~~~~~
+
+* script to update label
+
 1.2.0 (2019-01-16)
 ------------------